Modularize scatter plot logging
