+++ app/javascript/components/visualization/ScatterPlot.js
@@ -5,10 +5,10 @@ import _uniqueId from 'lodash/uniqueId'
 import Plotly from 'plotly.js-dist'
 
 import { fetchCluster } from 'lib/scp-api'
+import { logScatterPlot } from 'lib/scp-api-metrics'
 import { labelFont, getColorBrewerColor } from 'lib/plot'
 import { useUpdateEffect } from 'hooks/useUpdate'
 import PlotTitle from './PlotTitle'
-import { log } from 'lib/metrics-api'
 import useErrorMessage, { checkScpApiResponse } from 'lib/error-message'
 import { withErrorBoundary } from 'lib/ErrorBoundary'
 
@@ -64,27 +64,10 @@ function RawScatterPlot({
 
       Plotly.react(graphElementId, scatter.data, layout)
 
-      const perfTimeFrontend = performance.now() - perfTimeFrontendStart
+      const perfTimes = { perfTime, perfTimeFrontendStart }
+      const plotData = { scatter, genes, width, height }
 
-      const perfTimeFull = perfTime + perfTimeFrontend
-
-      const perfLogProps = {
-        'perfTime:backend': perfTime, // Time for API call
-        'perfTime:frontend': Math.round(perfTimeFrontend), // Time from API call *end* to plot render end
-        'perfTime': Math.round(perfTimeFull), // Time from API call *start* to plot render end,
-        'numPoints': scatter.numPoints, // How many cells are we plotting?
-        genes,
-        'gene': scatter.gene,
-        'is3D': scatter.is3D,
-        'layout:width': width, // Pixel width of graph
-        'layout:height': height, // Pixel height of graph
-        'numAnnotSelections': scatter.annotParams.values.length,
-        'annotName': scatter.annotParams.name,
-        'annotType': scatter.annotParams.type,
-        'annotScope': scatter.annotParams.scope
-      }
-
-      log('plot:scatter', perfLogProps)
+      logScatterPlot(perfTimes, plotData)
 
       setClusterData(scatter)
       setShowError(false)
